+++ core/data/src/main/java/com/hegunhee/data/di/RepositoryModule.kt
@@ -4,10 +4,12 @@ import com.hegunhee.data.repository.DefaultRepository
 import com.hegunhee.data.repository.twitch.DefaultTwitchGameStreamRepository
 import com.hegunhee.data.repository.twitch.DefaultTwitchSearchStreamerRepository
 import com.hegunhee.data.repository.twitch.DefaultTwitchStreamRepository
+import com.hegunhee.data.repository.twitch.DefaultTwitchUpdateStreamerRepository
 import com.hegunhee.domain.repository.Repository
 import com.hegunhee.domain.repository.twitch.TwitchGameStreamRepository
 import com.hegunhee.domain.repository.twitch.TwitchSearchStreamerRepository
 import com.hegunhee.domain.repository.twitch.TwitchStreamRepository
+import com.hegunhee.domain.repository.twitch.TwitchUpdateStreamerRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -33,4 +35,9 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun provideTwitchSearchStreamerDataRepository(defaultTwitchSearchStreamerRepository: DefaultTwitchSearchStreamerRepository) : TwitchSearchStreamerRepository
+
+    @Binds
+    @Singleton
+    abstract fun provideTwitchUpdateStreamerRepository(defaultTwitchUpdateStreamerRepository: DefaultTwitchUpdateStreamerRepository): TwitchUpdateStreamerRepository
+
 }
