+++ core/data/src/test/java/com/hegunhee/data/repository/twitch/InsertDeleteDataTest.kt
@@ -3,7 +3,6 @@ package com.hegunhee.data.repository.twitch
 import com.hegunhee.data.dataSource.local.LocalDataSource
 import com.hegunhee.data.dataSource.remote.RemoteDataSource
 import com.hegunhee.data.mapper.toStreamerEntity
-import com.hegunhee.data.repository.DefaultRepository
 import com.hegunhee.domain.model.twitch.StreamerData
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
@@ -19,7 +18,7 @@ import org.mockito.kotlin.whenever
 class InsertDeleteDataTest {
 
     @InjectMocks
-    private lateinit var sut: DefaultRepository
+    private lateinit var sut: DefaultTwitchUpdateStreamerRepository
 
     @Mock
     private lateinit var remoteDataSource: RemoteDataSource
